fix(*): change the timeout mechanism of high-watermark
if we use a newView timer to track the update of low-watermark, we need to stop a newView timer when
normal checkpoint has finished, but sometimes, we cannot just stop a newView timer at stable
checkpoint as a normal checkpoint mechanism can only be used for garbage collection and a newView
timer could be started by vc or any other process to track the node's behavior. so that we should
set a specific timer for high-watermark and trigger view-change if it expires, and here the aim of
view-change process is to find a correct stable checkpoint

fix #flato-2834

fix(*): make high-watermark timer legal

we know that a quorum set of nodes have already executed a block of high-watermark, but some of them missed the
checkpoint message. now, we can wait for another nodes' execution or just trigger view-change. as the view-change
process may take too long time to harm the throughput, we would like to wait for others' execution at first.
here, they should increase their watermark for a K-interval. as one block's consensus timeout is equal to
requestTimer, the high-watermark timer should be at least K*requestTimer. but checkpoint message should be sent
after executed, we have to take the latency of executor into thought. so that we would like to set high-watermark
timer 2*k*requestTimer